move getExpectationInfo to its own file
